requirements.txt
RPi.GPIO==0.5.11
pyserial==2.5
